backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0    # PostgreSQL adapter (optional; SQLite works without it)

# ── Auth & Security ───────────────────────────────────────────────────────────
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6   # Form data parsing (needed for file uploads)

# ── Data & ML ─────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
email-validator>=2.0.0
redis>=5.0.0              # Optional — used for caching when REDIS_URL is set

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
